Refactored Clients and Suppliers Endpoints. Relocated mapping logic from controller to service. Made some interfaces and abstract class to minimize duplicated code. Added constructor to select statement in JPQL. Optimized Get endpoints.

+++ Warehouse/src/main/java/com/example/warehouse/domain/dto/clientAndSupplierDtos/ClientSummaryDto.java
@@ -17,7 +17,7 @@ public class ClientSummaryDto {
     private String email;
     private String phoneNumber;
     private String address;
-    private Integer transactionsCount;
+    private Long transactionsCount;
 }
