+++ Warehouse/src/main/java/com/example/warehouse/domain/dto/clientAndSupplierDtos/SupplierSummaryDto.java
@@ -17,5 +17,5 @@ public class SupplierSummaryDto {
     private String email;
     private String phoneNumber;
     private String address;
-    private Integer transactionsCount;
+    private Long transactionsCount;
 }
